+++ IR/IR/InMemoryIndexBuilder.cpp
@@ -132,7 +132,7 @@ void InMemoryIndexBuilder::finalize(DocumentTable * documentTable)
 		outputFile.write((const char *)&iCompressor->getCompressorId(), sizeof(int));
 
 		// write the DocumentMetaDatas number in DocumentTable
-		outputFile.write((const char *)&documentTable->getDocumentNumber(), sizeof(unsigned long long));
+		outputFile.write((const char *)&documentTable->getDocumentNumber(), sizeof(unsigned int));
 
 		// write the DocumentMetaDatas
 		outputFile.write((const char *)documentTable->getFinalizedDocumentTable(), documentTable->getDocumentNumber()*sizeof(DocumentMetaData));
